requirements.txt
aiodns==3.0.0
aiofiles==0.7.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.4
black==23.10.0
blinker==1.6.3
certifi==2020.6.20
cffi==1.15.1
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.1.0
cleo==1.0.0a4
click==8.1.3
colorzero==1.1
crashtest==0.3.1
deezer-py==1.3.6
distlib==0.3.7
distro==1.5.0
filelock==3.12.2
flake8==6.1.0
flask==3.0.0
frozenlist==1.3.3
gpiozero==1.6.2
identify==2.5.30
idna==2.10
importlib-metadata==6.8.0
iso8601==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
m3u8==0.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mfrc522==0.0.7
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.11.2
pathvalidate==2.5.2
pigpio==1.78
Pillow==9.5.0
platformdirs==3.9.1
pre-commit==3.5.0
pycares==4.3.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodomex==3.18.0
pyflakes==3.1.0
pylev==1.4.0
python-apt==2.2.1
PyYAML==6.0.1
requests==2.25.1
RPi.GPIO==0.7.1
simple-term-menu==1.6.1
six==1.16.0
spidev==3.6
ssh-import-id==5.10
streamrip==1.9.7
toml==0.10.1
tomli==2.0.1
tomlkit==0.7.2
tqdm==4.65.0
typing-extensions==4.8.0
urllib3==1.26.5
virtualenv==20.24.1
werkzeug==3.0.0
yarl==1.9.2
zipp==3.17.0
